requirements.txt
git+git://github.com/haven-ai/haven.git@master#egg=haven
torchmeta==1.4.5
